dev-java/javacup: do not depend on jflex, build with bundled JFlex.jar
There is a circular dependency between javacup and jflex and in the past
we first built jflex using a bundled javacup-runtime. The newer version
of jflex-1.9.1 however does not provide such bundled javacup-runtime.

For this reason we change the build order and first build javacup using
the bundled bin/JFlex.jar and add "dev-java/jflex" to BDEPEND.

Signed-off-by: Volkmar W. Pogatzki <gentoo@pogatzki.net>