dev-java/javacup: do not depend on jflex, build with bundled JFlex.jar
There is a circular dependency between javacup and jflex and in the past
we used to build jflex with a bundled javacup-runtime. The newer version
of jflex-1.9.1 however does not provide such bundled javacup-runtime.

For this reason we change the emege order to build javacup first using
its bundled bin/JFlex.jar.

Also, PORTAGE_QUIET=1 is re-added to allow re-compilation with itself
without receiving 'QA Notice: java-pkg_ensure-dep' (bug #937047).

Bug: https://bugs.gentoo.org/937047
Signed-off-by: Volkmar W. Pogatzki <gentoo@pogatzki.net>