dev-java/java-gentoo-target: new package, add 1.8
All packages which inherit java-utils-2.eclass have an automatic
dependency on java-gentoo-target. It installs a file setting the eclass
variables JAVA_GENTOO_TARGET and JAVA_GENTOO_SOURCE.

These variables are parsed by the eclass functions java-pkg_get-target
and java-pkg_get-source.

SLOT must always be "0" and the subslot according to the lowest javac
target value actually supported by Gentoo.

The idea is to update to version 9 with SLOT="0/9" when Gentoo stops
supporting 1.8 in order to trigger recompilation of all packages which
were compiled for the old target.

Signed-off-by: Volkmar W. Pogatzki <gentoo@pogatzki.net>