java-pkg-simple.eclass: new function java-pkg-simple_generate-module-info
There are packages where module-info.java is not provided in the sources
but needs to be generated by the build system. We use the built-in jdeps
function with the --generate-module-info option which became available with
Java 11.

It generates the module-info.java file based on an intermediate jar file
which needs to be named as "${JAVA_MODULE_NAME}.jar" and places it in
the "${JAVA_MODULES_DIR}/${JAVA_MODULE_NAME}/versions/9" directory.

For this purpose we add two new eclass variables:
- JAVA_MODULE_NAME
- JAVA_MODULES_DIR

When both JAVA_MODULE_NAME and JAVA_MODULES_DIR are defined in the
ebuild we
- compile the sources still without module-info
- package them as an intermediate {JAVA_MODULE_NAME}.jar
- let java-pkg-simple_generate-module-info generate the module-info
- compile module-info.java with the --patch-module option
- package the final jar file with the module-info.class in a directory
  specific for --release 9.

Signed-off-by: Volkmar W. Pogatzki <gentoo@pogatzki.net>