dev-java/jboss-logmanager: new package, add 2.1.9
This package was removed on 2021-05-26 with commit #5494b20eb4.
Version 2.1.9 is a dependency of jboss-vfs which is needed for updating
dev-java/reflections to 0.10.2. It uses the new eclass functionality for
multi-release compilation but at least one class in META-INF/versions/9/
fails validation by java packager. This is worked-around by putting
JAVA_PACKAGER_ZIP="9" which triggers another new eclass functionality
using app-arch/zip instead of the java packager.

Signed-off-by: Volkmar W. Pogatzki <gentoo@pogatzki.net>