dev-db/mysql-workbench: Allow to build with system-vm set to openjdk-8
This ebuild did not inherit a java eclass nor did it depend on
virtual/jdk.  On systems having java system-vm set to openjdk{,-bin}-8
this was causing errors.

This commit adds java-pkg-2 to inherit, switches dependencies to java
defaults and adds the pkg_setup() section with java-pkg-2_pkg_setup.

Revbump because of dependency changes.

Bug: https://bugs.gentoo.org/933844
Closes: https://bugs.gentoo.org/914052
Signed-off-by: Volkmar W. Pogatzki <gentoo@pogatzki.net>